+++ server/src/nest/feeds/feeds.module.ts
@@ -1,10 +1,13 @@
 import { Module } from '@nestjs/common';
-import { TripsModule } from '../trips/trips.module';
+import { CalendarModule } from '../calendar/calendar.module';
+import { DatabaseModule } from '../database/database.module';
 import { FeedsService } from './feeds.service';
 import { FeedsPublicController, TripFeedTokenController, UserFeedTokenController } from './feeds.controller';
 
 @Module({
-  imports: [TripsModule],
+  // Calendars, not the trip aggregate: feeds only ever needed an ICS string, and
+  // importing TripsModule for it pulled budget, packing, places and the rest in.
+  imports: [CalendarModule, DatabaseModule],
   controllers: [FeedsPublicController, TripFeedTokenController, UserFeedTokenController],
   providers: [FeedsService],
 })
