+++ server/tests/e2e/feeds.e2e.test.ts
@@ -42,12 +42,12 @@ const SAMPLE_ICS =
   'BEGIN:VEVENT\r\nUID:trek-trip-x@trek\r\nDTSTAMP:20260101T000000Z\r\n' +
   'DTSTART;VALUE=DATE:20260101\r\nDTEND;VALUE=DATE:20260102\r\nSUMMARY:Sample\r\nEND:VEVENT\r\n' +
   'END:VCALENDAR\r\n';
-// Since the trip fold, FeedsService injects TripsService — the mock is a spy on
-// the container singleton (created in beforeAll, after build()).
+// FeedsService injects CalendarService — the mock is a spy on the container
+// singleton (created in beforeAll, after build()).
 const exportICS = vi.fn();
 
 import { FeedsModule } from '../../src/nest/feeds/feeds.module';
-import { TripsService } from '../../src/nest/trips/trips.service';
+import { CalendarService } from '../../src/nest/calendar/calendar.service';
 import { RealtimeModule } from '../../src/nest/realtime/realtime.module';
 import { TrekExceptionFilter } from '../../src/nest/common/trek-exception.filter';
 
@@ -73,7 +73,7 @@ describe('Calendar-feed e2e (real auth guard + temp SQLite)', () => {
     seedUser(db as never, { id: 1, username: 'e2e-user' });
     seedUser(db as never, { id: 2, username: 'other-user', email: 'other@example.test' });
     app = await build();
-    vi.spyOn(app.get(TripsService), 'exportICS').mockImplementation(exportICS as never);
+    vi.spyOn(app.get(CalendarService), 'exportICS').mockImplementation(exportICS as never);
     server = app.getHttpServer();
   });
